requirement.txt
astroid==2.3.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-friendship==1.8.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
pycryptodome==3.9.7
pylint==2.4.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.8
wrapt==1.11.2
